+++ resources/views/cp/space.blade.php
@@ -1,0 +1,64 @@
+{{-- Same seeded star field as the SocraNext AgentSpace component. --}}
+<div class="sncp-space-decor" aria-hidden="true">
+<span class="sncp-space__nebula sncp-space__nebula--1"></span>
+<span class="sncp-space__nebula sncp-space__nebula--2"></span>
+<span class="sncp-space__nebula sncp-space__nebula--3"></span>
+<div class="sncp-space__planet sncp-space__planet--tr"><div class="sncp-space__planet-ring"></div></div>
+<span class="sncp-space__star" style="top:34.0439%;left:18.7726%;width:2.250px;height:2.250px;animation-duration:3.960s;animation-delay:1.671s"></span>
+<span class="sncp-space__star" style="top:8.5475%;left:85.3743%;width:2.264px;height:2.264px;animation-duration:2.535s;animation-delay:3.464s"></span>
+<span class="sncp-space__star" style="top:82.6087%;left:19.9030%;width:1.564px;height:1.564px;animation-duration:4.317s;animation-delay:1.987s"></span>
+<span class="sncp-space__star" style="top:50.1131%;left:50.7520%;width:2.033px;height:2.033px;animation-duration:2.284s;animation-delay:0.355s"></span>
+<span class="sncp-space__star" style="top:16.6601%;left:10.1917%;width:2.870px;height:2.870px;animation-duration:3.741s;animation-delay:3.030s"></span>
+<span class="sncp-space__star" style="top:25.3502%;left:34.6274%;width:1.305px;height:1.305px;animation-duration:4.792s;animation-delay:0.909s"></span>
+<span class="sncp-space__star" style="top:62.8314%;left:98.2624%;width:2.217px;height:2.217px;animation-duration:3.338s;animation-delay:3.719s"></span>
+<span class="sncp-space__star" style="top:80.5194%;left:20.0031%;width:2.135px;height:2.135px;animation-duration:2.993s;animation-delay:2.829s"></span>
+<span class="sncp-space__star" style="top:8.4143%;left:23.0496%;width:2.070px;height:2.070px;animation-duration:4.668s;animation-delay:0.868s"></span>
+<span class="sncp-space__star" style="top:81.7961%;left:9.5570%;width:2.290px;height:2.290px;animation-duration:2.288s;animation-delay:1.663s"></span>
+<span class="sncp-space__star" style="top:85.1195%;left:97.5218%;width:2.390px;height:2.390px;animation-duration:3.799s;animation-delay:3.838s"></span>
+<span class="sncp-space__star" style="top:98.6729%;left:34.7088%;width:2.003px;height:2.003px;animation-duration:4.653s;animation-delay:1.441s"></span>
+<span class="sncp-space__star" style="top:85.4396%;left:73.2279%;width:1.624px;height:1.624px;animation-duration:3.840s;animation-delay:1.295s"></span>
+<span class="sncp-space__star is-bright" style="top:95.3205%;left:54.8240%;width:1.878px;height:1.878px;animation-duration:3.140s;animation-delay:3.836s"></span>
+<span class="sncp-space__star is-bright" style="top:26.3137%;left:6.0404%;width:1.121px;height:1.121px;animation-duration:3.433s;animation-delay:0.820s"></span>
+<span class="sncp-space__star" style="top:79.8922%;left:99.4280%;width:2.868px;height:2.868px;animation-duration:4.425s;animation-delay:3.851s"></span>
+<span class="sncp-space__star" style="top:75.1076%;left:53.0658%;width:1.141px;height:1.141px;animation-duration:4.071s;animation-delay:1.525s"></span>
+<span class="sncp-space__star" style="top:83.0258%;left:32.4484%;width:1.870px;height:1.870px;animation-duration:3.386s;animation-delay:3.633s"></span>
+<span class="sncp-space__star" style="top:24.3885%;left:33.2154%;width:2.858px;height:2.858px;animation-duration:3.757s;animation-delay:1.946s"></span>
+<span class="sncp-space__star" style="top:66.7456%;left:75.7557%;width:2.129px;height:2.129px;animation-duration:4.911s;animation-delay:1.329s"></span>
+<span class="sncp-space__star" style="top:67.8667%;left:34.8740%;width:2.123px;height:2.123px;animation-duration:3.936s;animation-delay:2.774s"></span>
+<span class="sncp-space__star is-bright" style="top:17.0500%;left:37.4523%;width:2.287px;height:2.287px;animation-duration:4.067s;animation-delay:3.285s"></span>
+<span class="sncp-space__star" style="top:53.7799%;left:16.0955%;width:2.247px;height:2.247px;animation-duration:2.609s;animation-delay:0.556s"></span>
+<span class="sncp-space__star" style="top:5.5536%;left:45.6091%;width:2.723px;height:2.723px;animation-duration:2.520s;animation-delay:3.616s"></span>
+<span class="sncp-space__star" style="top:4.3494%;left:74.5869%;width:1.446px;height:1.446px;animation-duration:4.184s;animation-delay:2.115s"></span>
+<span class="sncp-space__star" style="top:96.3621%;left:54.9490%;width:1.047px;height:1.047px;animation-duration:2.785s;animation-delay:1.454s"></span>
+<span class="sncp-space__star is-bright" style="top:50.1706%;left:82.0574%;width:1.001px;height:1.001px;animation-duration:3.822s;animation-delay:0.633s"></span>
+<span class="sncp-space__star" style="top:94.4645%;left:13.6225%;width:2.556px;height:2.556px;animation-duration:4.203s;animation-delay:1.321s"></span>
+<span class="sncp-space__star" style="top:52.4981%;left:62.9287%;width:1.851px;height:1.851px;animation-duration:3.390s;animation-delay:0.510s"></span>
+<span class="sncp-space__star is-bright" style="top:47.6616%;left:80.6326%;width:1.453px;height:1.453px;animation-duration:3.214s;animation-delay:2.569s"></span>
+<span class="sncp-space__star" style="top:39.9481%;left:62.6423%;width:2.538px;height:2.538px;animation-duration:2.577s;animation-delay:2.940s"></span>
+<span class="sncp-space__star" style="top:54.4987%;left:84.6286%;width:1.665px;height:1.665px;animation-duration:3.368s;animation-delay:1.207s"></span>
+<span class="sncp-space__star" style="top:99.6188%;left:84.8428%;width:1.796px;height:1.796px;animation-duration:2.778s;animation-delay:3.965s"></span>
+<span class="sncp-space__star is-bright" style="top:46.5930%;left:60.1315%;width:1.749px;height:1.749px;animation-duration:3.459s;animation-delay:1.561s"></span>
+<span class="sncp-space__star" style="top:86.8644%;left:38.5789%;width:2.119px;height:2.119px;animation-duration:4.577s;animation-delay:1.360s"></span>
+<span class="sncp-space__star" style="top:27.2902%;left:71.6564%;width:1.390px;height:1.390px;animation-duration:2.258s;animation-delay:2.116s"></span>
+<span class="sncp-space__star" style="top:90.3315%;left:24.4843%;width:2.806px;height:2.806px;animation-duration:3.645s;animation-delay:2.550s"></span>
+<span class="sncp-space__star" style="top:62.1646%;left:55.8096%;width:1.661px;height:1.661px;animation-duration:3.778s;animation-delay:3.275s"></span>
+<span class="sncp-space__star" style="top:91.1686%;left:61.5769%;width:2.989px;height:2.989px;animation-duration:4.940s;animation-delay:0.626s"></span>
+<span class="sncp-space__star" style="top:80.4592%;left:94.4056%;width:2.828px;height:2.828px;animation-duration:3.195s;animation-delay:0.317s"></span>
+<span class="sncp-space__star" style="top:22.3439%;left:96.2694%;width:2.337px;height:2.337px;animation-duration:4.240s;animation-delay:3.694s"></span>
+<span class="sncp-space__star" style="top:47.9233%;left:30.3034%;width:2.179px;height:2.179px;animation-duration:2.928s;animation-delay:2.038s"></span>
+<span class="sncp-space__star" style="top:81.8245%;left:25.5742%;width:2.361px;height:2.361px;animation-duration:4.874s;animation-delay:1.955s"></span>
+<span class="sncp-space__star" style="top:71.5709%;left:15.5572%;width:1.904px;height:1.904px;animation-duration:3.026s;animation-delay:1.003s"></span>
+<span class="sncp-space__star" style="top:51.0099%;left:31.8807%;width:1.892px;height:1.892px;animation-duration:4.259s;animation-delay:0.717s"></span>
+<span class="sncp-space__star" style="top:40.2814%;left:81.7711%;width:2.839px;height:2.839px;animation-duration:3.896s;animation-delay:1.551s"></span>
+<span class="sncp-space__star is-bright" style="top:9.5854%;left:35.4665%;width:2.518px;height:2.518px;animation-duration:4.455s;animation-delay:2.779s"></span>
+<span class="sncp-space__star is-bright" style="top:59.9625%;left:95.9388%;width:2.950px;height:2.950px;animation-duration:3.647s;animation-delay:2.006s"></span>
+<span class="sncp-space__star is-bright" style="top:88.4436%;left:36.0235%;width:1.694px;height:1.694px;animation-duration:3.782s;animation-delay:1.248s"></span>
+<span class="sncp-space__star is-bright" style="top:61.3415%;left:32.3841%;width:2.465px;height:2.465px;animation-duration:2.477s;animation-delay:1.755s"></span>
+<span class="sncp-space__star" style="top:37.9964%;left:95.8989%;width:2.096px;height:2.096px;animation-duration:4.557s;animation-delay:0.243s"></span>
+<span class="sncp-space__star" style="top:47.7577%;left:20.4957%;width:1.121px;height:1.121px;animation-duration:4.834s;animation-delay:3.735s"></span>
+<span class="sncp-space__star" style="top:80.8247%;left:96.2856%;width:2.989px;height:2.989px;animation-duration:4.002s;animation-delay:0.282s"></span>
+<span class="sncp-space__star" style="top:69.0287%;left:38.1146%;width:2.061px;height:2.061px;animation-duration:2.921s;animation-delay:1.739s"></span>
+<span class="sncp-space__star" style="top:46.1961%;left:90.8179%;width:1.649px;height:1.649px;animation-duration:4.069s;animation-delay:3.716s"></span>
+<span class="sncp-space__star" style="top:58.5633%;left:48.6521%;width:1.664px;height:1.664px;animation-duration:2.530s;animation-delay:2.259s"></span>
+<span class="sncp-space__shoot sncp-space__shoot--1"></span><span class="sncp-space__shoot sncp-space__shoot--2"></span>
+</div>